Add files via upload
Trade closing is prototyped. For the same Client, Stock and Banking platform, there might be more than single entries bought at different date and thus when a certain quantity is requested to be closed, we also require comparison to the existing entries. The previous entries are removed as per FIFO routine.